adds mocha scripts & renames tests folder
